pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Mamadou",email = "kmma.960@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)"
]
